DotGameServer/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
Faker==4.0.1
httpie==2.0.0
idna==2.9
Pillow==7.0.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.8
wincertstore==0.2
